feat: Add support for troubleshooting screen
This troubleshooting section (which can be accessed by clicking the
title of Force Refresh three times) gives users access to debugging
information that can be used when seeking support. Currently, this
section can be used to turn on debugging mode, which will allow logs
to be output in the browser console.

In the future, this section could include additional helpful tools, like
a console, to help users diagnose issues.