fix: Remove nonce for client-side calls
Fixes an issue when the client-side nonce would become cached and cause
calls to the WordPress API to fail. Since visitors to the website are
not always logged into WordPress, using the nonce isn't required from
a security standpoint (since API data are not unique to any single
visitor and the nonce would be generalized amongst all users).